+++ src/test/java/com/joel/flightreservations/domain/model/flight/FlightTest.java
@@ -13,26 +13,26 @@ public class FlightTest {
     public void testReserveSeats() throws Exception {
         Flight f = new Flight(SampleFlights.BA149);
         f.reserveSeats(10, 0);
-        Integer seats = 80;
+        Long seats = 80L;
         assertEquals(seats, f.getEconomyClassVacancies());
         f.unreserveSeats(10, 0);
-        seats = 90;
+        seats = 90L;
         assertEquals(seats, f.getEconomyClassVacancies());
         f.reserveSeats(90, 0);
-        seats = 0;
+        seats = 0L;
         assertEquals(seats, f.getEconomyClassVacancies());
         f.unreserveSeats(90, 0);
         f.reserveSeats(0, 10);
-        seats = 0;
+        seats = 0L;
         assertEquals(seats, f.getBusinessClassVacancies());
         f.unreserveSeats(0, 10);
         f.reserveSeats(0, 8);
-        seats = 2;
+        seats = 2L;
         assertEquals(seats, f.getBusinessClassVacancies());
         f.unreserveSeats(0, 8);
         f.reserveSeats(10, 5);
-        seats = 80;
-        Integer seats1 = 5;
+        seats = 80L;
+        Long seats1 = 5L;
         assertEquals(seats, f.getEconomyClassVacancies());
         assertEquals(seats1, f.getBusinessClassVacancies());
     }
